MS-128: Move containers up for consistency
