+++ frontend/src/containers/Home.jsx
@@ -1,5 +1,5 @@
-import Button from '../../components/Button';
-import Header from '../../components/Header';
+import Button from '../components/Button';
+import Header from '../components/Header';
 
 export default function Home() {
   return (
